# Refactor `scope` commands