--- a/crates/nu-cmd-lang/src/core_commands/scope/aliases.rs
+++ b/crates/nu-cmd-lang/src/core_commands/scope/aliases.rs
@@ -35,7 +35,7 @@ impl Command for ScopeAliases {
         let ctrlc = engine_state.ctrlc.clone();
 
         let mut scope_data = ScopeData::new(engine_state, stack);
-        scope_data.populate_all();
+        scope_data.populate_decls();
 
         Ok(scope_data.collect_aliases(span).into_pipeline_data(ctrlc))
     }