--- a/crates/nu-cmd-lang/src/core_commands/scope/commands.rs
+++ b/crates/nu-cmd-lang/src/core_commands/scope/commands.rs
@@ -35,7 +35,7 @@ impl Command for ScopeCommands {
         let ctrlc = engine_state.ctrlc.clone();
 
         let mut scope_data = ScopeData::new(engine_state, stack);
-        scope_data.populate_all();
+        scope_data.populate_decls();
 
         Ok(scope_data.collect_commands(span).into_pipeline_data(ctrlc))
     }