--- a/crates/nu-cmd-lang/src/core_commands/scope/mod.rs
+++ b/crates/nu-cmd-lang/src/core_commands/scope/mod.rs
@@ -2,6 +2,7 @@ mod aliases;
 mod command;
 mod commands;
 mod engine_stats;
+mod externs;
 mod modules;
 mod variables;
 
@@ -9,5 +10,6 @@ pub use aliases::*;
 pub use command::*;
 pub use commands::*;
 pub use engine_stats::*;
+pub use externs::*;
 pub use modules::*;
 pub use variables::*;