--- a/crates/nu-command/src/help/help_externs.rs
+++ b/crates/nu-command/src/help/help_externs.rs
@@ -144,7 +144,7 @@ pub fn help_externs(
 
 fn build_help_externs(engine_state: &EngineState, stack: &Stack, span: Span) -> Vec<Value> {
     let mut scope = ScopeData::new(engine_state, stack);
-    scope.populate_all();
+    scope.populate_decls();
     scope.collect_externs(span)
 }