… crates/nu-command/tests/commands/help.rs
… crates/nu-parser/src/parse_keywords.rs
… crates/nu-protocol/src/alias.rs
… tests/scope/mod.rs