<!--
if this PR closes one or more issues, you can automatically link the PR with
them by using one of the [*linking keywords*](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword), e.g.
- this PR should close #xxxx
- fixes #xxxx

you can also mention related issues, PRs or discussions!
-->

# Description
<!--
Thank you for improving Nushell. Please, check our [contributing guide](../CONTRIBUTING.md) and talk to the core team before making major changes.

Description of your pull request goes here. **Provide examples and/or screenshots** if your changes affect the user experience.
-->

See "User-Facing Changes". Apart from that, it cleans up the code of `scope` a bit and fixes/re-enables some tests. Since `help` uses a similar code as `scope` in most places, most of the improvements can be observed in `help`. I did not tweak `help` further since it is expected `std help` will replace the built-in one anyway.

### How to use the IDs

All definitions in `scope` commands now have their ID. It could be used to perform tasks such as looking up which module contains a given command. Some examples below:

```nushell
# finder.nu

# Find which module a command belongs to
export def find-module [
    cmd_name: string  # name of the command to look up
] {
    let decl_id = try {
        scope commands | where name == $cmd_name | get 0.decl_id
    } catch {
        let span = (metadata $cmd_name).span
        error make {
            msg: $'Command "($cmd_name)" not found'
            label: {
                text: 'could not find command'
                start: $span.start
                end: $span.end
            }
        }
    }

    for row in (scope modules) {
        for cmd in $row.commands {
           if $cmd.decl_id == $decl_id {
               return {
                   name: $row.name            # name of the module
                   module_id: $row.module_id  # module's ID
               }
           }
        }
    }
}

# Find which commands of a module are in active use
#
# Does not recurse to submodules!
export def find-used-commands [
    module_name: string  # name of the module to look up
] {
    let module_commands = try {
        scope modules | where name == $module_name | get 0.commands
    } catch {
        let span = (metadata $module_name).span
        error make {
            msg: $'Module "($module_name)" not found'
            label: {
                text: 'could not find module'
                start: $span.start
                end: $span.end
            }
        }
    }

    mut res = []

    for cmd in $module_commands {
        for row in (scope commands) {
            if $cmd.decl_id == $row.decl_id {
                $res ++= [
                    {
                        used_name: $row.name   # imported name in the current scope
                        name: $cmd.name        # name defined in the module
                        decl_id: $row.decl_id  # command's ID
                    }
                ]
            }
        }
    }

    $res
}
```

```
> use finder.nu *

> find-module banner
 name        std
 module_id   11

> find-used-commands std
 #   used_name    name    decl_id
──────────────────────────────────
 0   banner      banner       424
```

### Benchmarks

Benchmark script
```nushell
# bench_scope.nu

use std bench

{
    aliases      : (bench { scope aliases } | get mean)
    commands     : (bench { scope commands } | get mean)
    engine-stats : (bench { scope engine-stats } | get mean)
    modules      : (bench { scope modules } | get mean)
    variables    : (bench { scope variables } | get mean)
    externs      : (bench { scope externs } | get mean)
}
```


## Performance

Results on my machine with my config. This will be different for each setup, as it depends on how many definitions you have.

| |aliases|commands|engine-stats|modules|variables|externs|
| - |-:|-:|-:|-:|-:|-:|
| baseline | 99µs 942ns|2ms 517µs 333ns|10µs 990ns|62µs 995ns|11µs 123ns||
| https://github.com/nushell/nushell/pull/10023/commits/9fff54c4ade9a1732da9115fee6d16485afff36b | 100µs 192ns|2ms 411µs 832ns|11µs 167ns|62µs 364ns|10µs 909ns||
| https://github.com/nushell/nushell/pull/10023/commits/7fdf5989faed681071acc6dcde140c5007e5736b |97µs 752ns|2ms 333µs 524ns|2µs 652ns|64µs 230ns|2µs 875ns||
| https://github.com/nushell/nushell/pull/10023/commits/2041043418eebccecaa1d850dcacfe9c3aed1ad0 |96µs 514ns|2ms 329µs 78ns|2µs 588ns|63µs 951ns|2µs 840ns||
| https://github.com/nushell/nushell/pull/10023/commits/c3f5037a6d463a43eadfabda06e3bf3365dc3042 |95µs 411ns|2ms 403µs 712ns|2µs 703ns|64µs 580ns|2µs 981ns||
| https://github.com/nushell/nushell/pull/10023/commits/96aeb9f11d9e3e6dc4519ecfa05c50696ba5bb30 |95µs 690ns|2ms 320µs 869ns|2µs 767ns|63µs 415ns|2µs 900ns|15µs 732ns|
| https://github.com/nushell/nushell/pull/10023/commits/0b2f26fa5191ecd7f477bede41ed0adcf0bb1d1c |100µs 598ns|2ms 348µs 30ns|2µs 719ns|121µs 299ns|2µs 918ns|15µs 880ns|
| https://github.com/nushell/nushell/pull/10023/commits/6c428883f66b239ae2e0b9662aa45e5d89f56720 |101µs 464ns|2ms 374µs 293ns|2µs 716ns|280µs 691ns|2µs 925ns|15µs 822ns|
| https://github.com/nushell/nushell/pull/10023/commits/c46c5ca958c7e20b1c21ddb6bfc1c6fc2f58f3c6 |103µs 920ns|2ms 378µs 838ns|2µs 786ns|285µs 213ns|2µs 982ns|16µs 319ns|

* `scope variables` is about 3.5x faster
* `scope modules` is about 4.5x slower (because it contains more stuff now)

Fixes https://github.com/nushell/nushell/issues/10006

# User-Facing Changes
<!-- List of all changes that impact the user experience here. This helps us keep track of breaking changes. -->

* (minor breaking change) Removes `module_name` field from `scope/help externs` commands (it seemed to be broken anyway).
* Fixes alias entries in `scope` and `help` commands.
* Adds `scope externs`.
* Adds ID for each definition in all `scope` commands.

# Tests + Formatting
<!--
Don't forget to add tests that cover your changes.

Make sure you've run and fixed any issues with these commands:

- `cargo fmt --all -- --check` to check standard code formatting (`cargo fmt --all` applies these changes)
- `cargo clippy --workspace -- -D warnings -D clippy::unwrap_used -A clippy::needless_collect -A clippy::result_large_err` to check that you're using the standard code style
- `cargo test --workspace` to check that all tests pass
- `cargo run -- -c "use std testing; testing run-tests --path crates/nu-std"` to run the tests for the standard library

> **Note**
> from `nushell` you can also use the `toolkit` as follows
> ```bash
> use toolkit.nu  # or use an `env_change` hook to activate it automatically
> toolkit check pr
> ```
-->

# After Submitting
<!-- If your PR had any user-facing changes, update [the documentation](https://github.com/nushell/nushell.github.io) after the PR is merged, if necessary. This will help us keep the docs up to date. -->