@fdncred 2023-08-16
I definitely have noticed that sometimes I would get duplicate aliases listed. I haven't paid much attention to it though nor do I know how to reproduce it. Hopefully this fixes that too.