@kubouch 2023-08-16
About the visibility, I've been contemplating removing `hide` which would remove the visibility concept as well.

We still have shadowing (caused by scopes and/or overlays) which I think is possible to track, but would be a little hassle to get. Not too bad, though. Currently `scope xxx` just flattens everything in one line of code, we'd have to do it a bit more manually to detect the shadowed symbols.