@fdncred 2023-08-16
Since you're already in here doing surgery, maybe it's a good time to add `is_const`?