@kubouch 2023-08-16
`is_const` is not merged yet, it's in https://github.com/nushell/nushell/pull/9499